XWayland: move cursor logic to it's own files
